Conflicts:
	platform/api/StampedAPI.py
	platform/servers/httpapi/v0/functions/comments.py

+++ platform/api/StampedAPI.py
@@ -544,7 +544,7 @@ class StampedAPI(AStampedAPI):
         for stamp in stamps:
             if stamp.credits is not None and len(stamp.credits) > 0:
                 for stampPreview in stamp.credits:
-                    self._stampDB.removeCredit(stampPreview.user.user_id, stamp)
+                    self._stampDB.removeCredit(stampPreview.user.user_id, stamp.stamp_id, stamp.user.user_id)
 
                     # Decrement user stats by one
                     self._userDB.updateUserStats(stampPreview.user.user_id, 'num_credits', increment=-1)
@@ -552,10 +552,12 @@ class StampedAPI(AStampedAPI):
             # Remove activity on stamp
             self._activityDB.removeActivityForStamp(stamp.stamp_id)
 
-        self._stampDB.removeStamps(stampIds)
+        if stampIds is not None and len(stampIds) > 0:
+            self._stampDB.removeStamps(stampIds)
+            self._stampStatsDB.removeStatsForStamps(stampIds)
+
         self._stampDB.removeAllUserStampReferences(account.user_id)
         self._stampDB.removeAllInboxStampReferences(account.user_id)
-        self._stampStatsDB.removeStatsForStamps(stampIds)
         ### TODO: If restamp, remove from credited stamps' comment list?
 
         # Remove comments
@@ -653,49 +655,6 @@ class StampedAPI(AStampedAPI):
             tasks.invoke(tasks.APITasks.updateProfileImage, args=[account.screen_name, fields['temp_image_url']])
 
         return self._accountDB.updateAccount(account)
-
-
-#    @API_CALL
-#    def updateAccountSettings(self, authUserId, data):
-#
-#        ### TODO: Reexamine how updates are done
-#        ### TODO: Verify that email address is unique, confirm it
-#
-#        account = self._accountDB.getAccount(authUserId)
-#
-#        old_screen_name = account.screen_name
-#
-#        # Import each item
-#        for k, v in data.iteritems():
-#            if k == 'password':
-#                v = convertPasswordForStorage(v)
-#            setattr(account, k, v)
-#
-#        ### TODO: Carve out "validate account" function
-#
-#        # Validate Screen Name
-#        account.screen_name = account.screen_name.strip()
-#        if not utils.validate_screen_name(account.screen_name):
-#            raise StampedInputError("Invalid format for screen name")
-#
-#        # Check blacklist
-#        if account.screen_name.lower() in Blacklist.screen_names:
-#            raise StampedInputError("Blacklisted screen name")
-#
-#        # Validate email address
-#        if account.email is not None:
-#            account.email = str(account.email).lower().strip()
-#            if not utils.validate_email(account.email):
-#                raise StampedInputError("Invalid format for email address")
-#
-#        self._accountDB.updateAccount(account)
-#
-#        # Asynchronously update profile picture link if screen name has changed
-#        if account.screen_name.lower() != old_screen_name.lower():
-#            tasks.invoke(tasks.APITasks.changeProfileImageNameAsync, args=[
-#                         old_screen_name.lower(), account.screen_name.lower()])
-
-        return account
 
     @API_CALL
     def changeProfileImageNameAsync(self, old_screen_name, new_screen_name):
@@ -1001,8 +960,6 @@ class StampedAPI(AStampedAPI):
             self._addLinkedFriendActivity(authUserId, 'facebook', userIds,
                                           body = 'Your Facebook friend %s joined Stamped.' % account.linked.facebook.linked_name)
             self._accountDB.addLinkedAccountAlertHistory(authUserId, 'facebook', account.linked.facebook.linked_user_id)
-
-
 
     @API_CALL
     def addToNetflixInstant(self, nf_user_id, nf_token, nf_secret, netflixId):
@@ -1612,10 +1569,6 @@ class StampedAPI(AStampedAPI):
                        authUserId=None,
                        category=None):
 
-        #entities = self._oldEntitySearch.searchEntities(query,
-        #                                                limit=10,
-        #                                                coords=coords,
-        #                                                category=category)
         coordsAsTuple = None if coords is None else (coords.lat, coords.lng)
         entities = self._newEntitySearch.searchEntities(category, query, limit=10, coords=coordsAsTuple)
 
@@ -1691,6 +1644,20 @@ class StampedAPI(AStampedAPI):
                                                             subcategory=subcategory,
                                                             coords=coordinates,
                                                             limit=limit)
+
+    @API_CALL
+    def getMenu(self, entityId):
+        menu = self._menuDB.getMenu(entityId)
+        if menu is None:
+            try:
+                self.mergeEntityId(entityId)
+                menu = self._menuDB.getMenu(entityId)
+            except Exception:
+                pass
+        if menu is None:
+            raise StampedMenuUnavailableError()
+        else:
+            return menu
 
     @API_CALL
     def completeAction(self, authUserId, **kwargs):
@@ -1786,7 +1753,12 @@ class StampedAPI(AStampedAPI):
         return stampedby
 
     def updateEntityStatsAsync(self, entityId):
-        entity = self._entityDB.getEntity(entityId)
+        try:
+            entity = self._entityDB.getEntity(entityId)
+        except StampedDocumentNotFoundError:
+            logs.warning("Entity not found: %s" % entityId)
+            return 
+
         if entity.sources.tombstone_id is not None:
             # Call async process to update references
             tasks.invoke(tasks.APITasks.updateTombstonedEntityReferences, args=[entity.entity_id])
@@ -2272,28 +2244,13 @@ class StampedAPI(AStampedAPI):
         if blurbData is not None:
             content.blurb = blurbData.strip()
 
-        # Add image to stamp
-        if imageUrl is not None:
-            ### TODO: Verify external image exists via head request?
-            """
-            # TODO:
-            response = utils.getHeadRequest(image_url)
-
-            if response is None:
-                raise StampedInputError("unable to access specified stamp image '%s'" % image_url)
-            else:
-                content_type = response.info().getheader('Content-Type')
-
-                if content_type != "image/jpeg":
-                    raise StampedInputError("invalid external image format; content-type must be image/jpeg")
-            """
-            pass
+        ### TODO: Verify external image exists via head request?
 
         # Create the stamp or get the stamp object so we can use its stamp_id for image filenames
         if stampExists:
-            stamp                       = self._stampDB.getStampFromUserEntity(user.user_id, entity.entity_id)
+            stamp = self._stampDB.getStampFromUserEntity(user.user_id, entity.entity_id)
         else:
-            stamp                       = Stamp()
+            stamp = Stamp()
 
         logs.debug('### addStamp section 1: %s' % (time.time() - t1))
         t1 = time.time()
@@ -2426,24 +2383,12 @@ class StampedAPI(AStampedAPI):
                 if item.user.user_id == authUserId:
                     continue
 
-                """
-                # Check if block exists between user and credited user
-
-                friendship              = Friendship()
-                friendship.user_id      = authUserId
-                friendship.friend_id    = item.user.user_id
-
-                if self._friendshipDB.blockExists(friendship) == True:
-                    logs.debug("Block exists")
-                    continue
-                """
-
                 # Check if the user has been given credit previously
                 if self._stampDB.checkCredit(item.user.user_id, stamp):
                     continue
 
                 # Assign credit
-                self._stampDB.giveCredit(item.user.user_id, stamp)
+                self._stampDB.giveCredit(item.user.user_id, stamp.stamp_id, stamp.user.user_id)
                 creditedUserIds.add(item.user.user_id)
 
                 # Add stats
@@ -2521,20 +2466,20 @@ class StampedAPI(AStampedAPI):
 
 
     @API_CALL
-    def shareStamp(self, authUserId, stampId, service_name, imageUrl):
+    def shareStamp(self, authUserId, stampId, serviceName, imageUrl):
         stamp = self.getStamp(stampId)
         account = self.getAccount(authUserId)
 
-        if service_name == 'facebook':
+        if serviceName == 'facebook':
             if account.linked is None or account.linked.facebook is None \
-               or account.linked.facebook.token is None or account.linked.facebook.linked_user_id is None:
+                or account.linked.facebook.token is None or account.linked.facebook.linked_user_id is None:
                 raise StampedLinkedAccountError('Cannot share stamp on facebook, missing necessary linked account information.')
             self._facebook.postToNewsFeed(account.linked.facebook.linked_user_id,
                                           account.linked.facebook.token,
                                           stamp.contents[-1].blurb,
                                           imageUrl)
         else:
-            raise StampedNoSharingForLinkedAccountError("Sharing is not implemented for service: %s" % service_name)
+            raise StampedNoSharingForLinkedAccountError("Sharing is not implemented for service: %s" % serviceName)
         return stamp
 
     @API_CALL
@@ -2543,8 +2488,11 @@ class StampedAPI(AStampedAPI):
 
     @API_CALL
     def removeStamp(self, authUserId, stampId):
-        stamp       = self._stampDB.getStamp(stampId)
-        stamp       = self._enrichStampObjects(stamp, authUserId=authUserId)
+        try:
+            stamp = self._stampDB.getStamp(stampId)
+        except StampedDocumentNotFoundError:
+            logs.info("Stamp has already been deleted")
+            return True
 
         # Verify user has permission to delete
         if stamp.user.user_id != authUserId:
@@ -2553,15 +2501,19 @@ class StampedAPI(AStampedAPI):
         # Remove stamp
         self._stampDB.removeStamp(stamp.stamp_id)
 
-        ### NOTE: we rely on deleted stamps remaining in user's inboxes to
-        # signal to clients that the stamp should be removed. this is not
-        # necessary for the user stamps collection
+        tasks.invoke(tasks.APITasks.removeStamp, args=[authUserId, stampId, stamp.entity.entity_id, stamp.credits])
+
+        return True
+
+    def removeStampAsync(self, authUserId, stampId, entityId, credits=None):
 
         # Remove from user collection
-        self._stampDB.removeUserStampReference(authUserId, stamp.stamp_id)
+        self._stampDB.removeUserStampReference(authUserId, stampId)
+
+        ### TODO: Remove from inboxes? Or let integrity checker do that?
 
         # Remove from stats
-        self._stampStatsDB.removeStampStats(stamp.stamp_id)
+        self._stampStatsDB.removeStampStats(stampId)
 
         ### TODO: Remove from activity? To do? Anything else?
 
@@ -2573,13 +2525,13 @@ class StampedAPI(AStampedAPI):
             self._commentDB.removeComment(commentId)
 
         # Remove activity
-        self._activityDB.removeActivityForStamp(stamp.stamp_id)
+        self._activityDB.removeActivityForStamp(stampId)
 
         # Remove as todo if necessary
         try:
-            self._todoDB.completeTodo(stamp.entity_id, authUserId, complete=False)
-        except Exception:
-            pass
+            self._todoDB.completeTodo(entityId, authUserId, complete=False)
+        except Exception as e:
+            logs.warning(e)
 
         ### TODO: Remove reference in other people's todos
 
@@ -2587,43 +2539,31 @@ class StampedAPI(AStampedAPI):
         ### TODO: Do an actual count / update?
         self._userDB.updateUserStats(authUserId, 'num_stamps',      increment=-1)
         self._userDB.updateUserStats(authUserId, 'num_stamps_left', increment=1)
+
         distribution = self._getUserStampDistribution(authUserId)
         self._userDB.updateDistribution(authUserId, distribution)
 
         # Update credit stats if credit given
-        if stamp.credits is not None and len(stamp.credits) > 0:
-            for item in stamp.credits:
+        if credits is not None and len(credits) > 0:
+            for item in credits:
                 # Only run if user is flagged as credited
                 if item.user.user_id is None:
                     continue
 
                 # Assign credit
-                self._stampDB.removeCredit(item.user.user_id, stamp)
+                self._stampDB.removeCredit(item.user.user_id, stampId, authUserId)
 
                 # Update credited user stats
                 self._userDB.updateUserStats(item.user.user_id, 'num_credits', increment=-1)
 
-        # Update modified timestamp
-        stamp.timestamp.modified = datetime.utcnow()
-
         # Update entity stats
-        tasks.invoke(tasks.APITasks.updateEntityStats, args=[stamp.entity.entity_id])
-
-        return stamp
+        tasks.invoke(tasks.APITasks.updateEntityStats, args=[entityId])
 
     @API_CALL
-    def getStamp(self, stampId, authUserId=None):
-        stamp       = self._stampDB.getStamp(stampId)
-        stamp       = self._enrichStampObjects(stamp, authUserId=authUserId)
-
-        # Check privacy of stamp
-        if stamp.user.user_id != authUserId and stamp.user.privacy == True:
-            friendship              = Friendship()
-            friendship.user_id      = user.user_id
-            friendship.friend_id    = authUserId
-
-            if not self._friendshipDB.checkFriendship(friendship):
-                raise StampedViewStampPermissionsError("Insufficient privileges to view stamp")
+    def getStamp(self, stampId, authUserId=None, enrich=True):
+        stamp = self._stampDB.getStamp(stampId)
+        if enrich:
+            stamp = self._enrichStampObjects(stamp, authUserId=authUserId)
 
         return stamp
 
@@ -2665,11 +2605,17 @@ class StampedAPI(AStampedAPI):
             return statsDict
 
     def updateStampStatsAsync(self, stampId):
+        try:
+            stamp = self._stampDB.getStamp(stampId)
+        except StampedDocumentNotFoundError:
+            logs.warning("Stamp not found: %s" % stampId)
+            ### TODO: Will returning None cause problems?
+            return None
+
         stats                   = StampStats()
         stats.stamp_id          = stampId
 
         MAX_PREVIEW             = 10
-        stamp                   = self._stampDB.getStamp(stampId)
         stats.last_stamped      = stamp.timestamp.stamped
 
         likes                   = self._stampDB.getStampLikes(stampId)
@@ -2921,10 +2867,6 @@ class StampedAPI(AStampedAPI):
         # Add activity for previous commenters
         ### TODO: Limit this to the last 20 comments or so
         for prevComment in self._commentDB.getCommentsForStamp(stamp.stamp_id):
-            # Skip if it was generated from a restamp
-            if prevComment.restamp_id:
-                continue
-
             repliedUserId = prevComment.user.user_id
 
             if repliedUserId not in commentedUserIds \
@@ -3863,7 +3805,14 @@ class StampedAPI(AStampedAPI):
 
     @API_CALL
     def addTodo(self, authUserId, entityRequest, stampId=None):
+        # Entity
         entity = self._getEntityFromRequest(entityRequest)
+
+        # User
+        user = self._userDB.getUser(authUserId).minimize()
+
+        # Friends
+        friendIds = self._friendshipDB.getFriends(user.user_id)
 
         todo                    = RawTodo()
         todo.entity             = entity.minimize()
@@ -3874,7 +3823,7 @@ class StampedAPI(AStampedAPI):
         if stampId is not None:
             todo.source_stamp_ids = [stampId]
 
-        # Check to verify that user hasn't already todoed entity
+        # Check to verify that user hasn't already todo'd entity
         try:
             testTodo = self._todoDB.getTodo(authUserId, entity.entity_id)
             if testTodo.todo_id is None:
@@ -3884,8 +3833,7 @@ class StampedAPI(AStampedAPI):
             exists = False
 
         if exists:
-            return testTodo
-            #raise StampedDuplicationError("Todo already exists")
+            return self._enrichTodo(testTodo, user=user, entity=entity, friendIds=friendIds, authUserId=authUserId)
 
         # Check if user has already stamped the todo entity, mark as complete and provide stamp_id, if so
         users_stamp = self._stampDB.getStampFromUserEntity(authUserId, entity.entity_id)
@@ -3904,15 +3852,11 @@ class StampedAPI(AStampedAPI):
         # Increment stats
         self._statsSink.increment('stamped.api.stamps.todos')
 
-        # User
-        user = self._userDB.getUser(authUserId).minimize()
 
         #stamp
         if stampId is not None:
             stamp = self._stampDB.getStamp(stampId)
             stampOwner = stamp.user.user_id
-
-        friendIds = self._friendshipDB.getFriends(user.user_id)
 
         # Enrich todo
         todo = self._enrichTodo(todo, user=user, entity=entity, stamp=users_stamp, friendIds=friendIds, authUserId=authUserId)
@@ -4334,29 +4278,6 @@ class StampedAPI(AStampedAPI):
 
 
     """
-    #     # ###### #     # #     #
-    ##   ## #      ##    # #     #
-    # # # # #      # #   # #     #
-    #  #  # ###### #  #  # #     #
-    #     # #      #   # # #     #
-    #     # #      #    ## #     #
-    #     # ###### #     #  #####
-    """
-
-    def getMenu(self, entityId):
-        menu = self._menuDB.getMenu(entityId)
-        if menu is None:
-            try:
-                self.mergeEntityId(entityId)
-                menu = self._menuDB.getMenu(entityId)
-            except Exception:
-                pass
-        if menu is None:
-            raise StampedMenuUnavailableError("Menu unavailable")
-        else:
-            return menu
-
-    """
     ######
     #     # ######  ####   ####  #      #    # ######
     #     # #      #      #    # #      #    # #
@@ -4675,51 +4596,6 @@ class StampedAPI(AStampedAPI):
     #       #   #  #  #  #  #    #   #   #
     #       #    # #   ##   #    #   #   ######
     """
-
-    def _addEntity(self, entity):
-        if entity is not None:
-            utils.log("[%s] adding 1 entity" % (self, ))
-
-            try:
-                #self._entityMatcher.addOne(entity)
-                entity2 = self._entityDB.addEntity(entity)
-
-                if 'place' in entity:
-                    self._placesEntityDB.addEntity(entity2)
-
-                return entity2
-            except Exception as e:
-                utils.log("[%s] error adding 1 entities:" % (self, ))
-                utils.printException()
-                # don't let error propagate
-
-    def _addEntities(self, entities):
-        entities = filter(lambda e: e is not None, entities)
-        numEntities = utils.count(entities)
-        if numEntities <= 0:
-            return
-
-        utils.log("[%s] adding %d entities" % (self, numEntities))
-
-        try:
-            entities2 = self._entityDB.addEntities(entities)
-            assert len(entities2) == len(entities)
-            place_entities = []
-
-            for i in xrange(len(entities2)):
-                entity = entities2[i]
-
-                if 'place' in entity:
-                    place_entities.append(entity)
-
-            if len(place_entities) > 0:
-                self._placesEntityDB.addEntities(place_entities)
-
-            return entities2
-        except Exception as e:
-            utils.log("[%s] error adding %d entities:" % (self, utils.count(entities)))
-            utils.printException()
-            # don't let error propagate
 
     def addClientLogsEntry(self, authUserId, entry):
         entry.user_id = authUserId
